Change parsing logic from parsing player names to parsing player ids
Origininally I was under the impression that some of the logs were malformed,
since some games end with ShutDownGame while others don't for instance. So
When I saw that sometimes multiple ClientUserInfoChanged had the same id
I assumed it was a malformed log, and not that it was possible to change your
name after connecting to a match. Upon further inspection I realized I was wrong,
and changed the parsing logic to store players by id.